+++ CHANGELOG.md
@@ -71,6 +71,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - Fixed `WindowLevel::Tooltip` on Windows platform ([#1737] by [@djeedai])
 - X11 backend now supports scaling([#1751] by [@Maan2003])
 - X11 backend now supports changing cursors ([#1755] by [@Maan2003])
+- X11 backend now uses the platform locale ([#1756] by [@Maan2003])
 
 ### Visual
 
@@ -710,6 +711,7 @@ Last release without a changelog :(
 [#1751]: https://github.com/linebender/druid/pull/1751
 [#1754]: https://github.com/linebender/druid/pull/1754
 [#1755]: https://github.com/linebender/druid/pull/1755
+[#1756]: https://github.com/linebender/druid/pull/1756
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
 [0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0